requirements.txt
streamlit>=1.28
diffusers
torch
transformers
accelerate
pillow
imageio
moviepy